# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountsResponse": {
        "properties": {
          "accounts": {
            "items": {
              "type": "string"
            },
            "title": "Accounts",
            "type": "array"
          },
          "max_accounts": {
            "title": "Max Accounts",
            "type": "integer"
          },
          "verified_accounts": {
            "items": {
              "type": "string"
            },
            "title": "Verified Accounts",
            "type": "array"
          }
        },
        "required": [
          "accounts",
          "verified_accounts",
          "max_accounts"
        ],
        "title": "AccountsResponse",
        "type": "object"
      },
      "AddAccountsRequest": {
        "properties": {
          "handles": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Handles",
            "type": "array"
          }
        },
        "required": [
          "handles"
        ],
        "title": "AddAccountsRequest",
        "type": "object"
      },
...
      "CreditState": {
        "properties": {
          "monthly_left": {
            "title": "Monthly Left",
            "type": "integer"
          },
          "monthly_limit": {
            "title": "Monthly Limit",
            "type": "integer"
          },
          "prepaid_left": {
            "title": "Prepaid Left",
            "type": "integer"
          },
          "prepaid_limit": {
            "title": "Prepaid Limit",
            "type": "integer"
          }
        },
        "required": [
          "monthly_left",
          "prepaid_left",
          "monthly_limit",
          "prepaid_limit"
        ],
        "title": "CreditState",
        "type": "object"
      },
...
      "GlobalSettings": {
        "properties": {
          "ai_prompt": {
            "default": "",
            "title": "Ai Prompt",
            "type": "string"
          },
          "newsletter_email": {
            "default": "",
            "title": "Newsletter Email",
            "type": "string"
          },
          "timezone": {
            "default": "America/New_York",
            "title": "Timezone",
            "type": "string"
          }
        },
        "title": "GlobalSettings",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImportAccountsRequest": {
        "properties": {
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "source"
        ],
        "title": "ImportAccountsRequest",
        "type": "object"
      },
      "ImportAccountsResponse": {
        "properties": {
          "accounts": {
            "items": {
              "type": "string"
            },
            "title": "Accounts",
            "type": "array"
          },
          "imported": {
            "items": {
              "type": "string"
            },
            "title": "Imported",
            "type": "array"
          },
          "skipped_due_to_limit": {
            "default": 0,
            "title": "Skipped Due To Limit",
            "type": "integer"
          }
        },
        "required": [
          "imported",
          "accounts"
        ],
        "title": "ImportAccountsResponse",
        "type": "object"
      },
      "MeResponse": {
        "properties": {
          "avatar": {
            "default": "",
            "title": "Avatar",
            "type": "string"
          },
          "created_at": {
            "default": "",
            "title": "Created At",
            "type": "string"
          },
          "credits": {
            "$ref": "#/components/schemas/CreditState"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "onboarded": {
            "default": false,
            "title": "Onboarded",
            "type": "boolean"
          },
          "onboarding_step": {
            "default": 0,
            "title": "Onboarding Step",
            "type": "integer"
          },
          "plan": {
            "title": "Plan",
            "type": "string"
          },
          "plan_info": {
            "$ref": "#/components/schemas/PlanInfo"
          },
          "tos_accepted": {
            "default": false,
            "title": "Tos Accepted",
            "type": "boolean"
          },
          "uid": {
            "title": "Uid",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageCounts"
          }
        },
        "required": [
          "uid",
          "email",
          "plan",
          "plan_info",
          "usage",
          "credits"
        ],
        "title": "MeResponse",
        "type": "object"
      },
      "Message": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Message",
        "type": "object"
      },
      "PlanInfo": {
        "properties": {
          "cancel_at_period_end": {
            "default": false,
            "title": "Cancel At Period End",
            "type": "boolean"
          },
          "current_period_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period End"
          },
          "current_period_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period Start"
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "PlanInfo",
        "type": "object"
      },
...
      "UpdateMeRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "onboarded": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarded"
          },
          "onboarding_step": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Step"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "tos_accepted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tos Accepted"
          }
        },
        "title": "UpdateMeRequest",
        "type": "object"
      },
      "UpdateSettingsRequest": {
        "properties": {
          "ai_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Prompt"
          },
          "newsletter_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Newsletter Email"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "title": "UpdateSettingsRequest",
        "type": "object"
      },
...
      "UsageCounts": {
        "properties": {
          "n_chat_messages": {
            "default": 0,
            "title": "N Chat Messages",
            "type": "integer"
          },
          "n_followers_scraped": {
            "default": 0,
            "title": "N Followers Scraped",
            "type": "integer"
          },
          "n_generations": {
            "default": 0,
            "title": "N Generations",
            "type": "integer"
          },
          "n_newsletters_sent": {
            "default": 0,
            "title": "N Newsletters Sent",
            "type": "integer"
          }
        },
        "title": "UsageCounts",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyAccountsResponse": {
        "properties": {
          "not_found": {
            "items": {
              "type": "string"
            },
            "title": "Not Found",
            "type": "array"
          },
          "verified_accounts": {
            "items": {
              "type": "string"
            },
            "title": "Verified Accounts",
            "type": "array"
          }
        },
        "required": [
          "verified_accounts",
          "not_found"
        ],
        "title": "VerifyAccountsResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API wrapping the FeedTLDR pipeline (scrape, summarize, narrate, email). Authenticated endpoints expect a Firebase ID token as a Bearer token.",
    "title": "FeedTLDR API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me": {
      "delete": {
        "description": "Delete the account: confirmation email, then Firebase Auth + Firestore\n(parity with the legacy profile delete flow).",
        "operationId": "delete_me_v1_me_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Me",
        "tags": [
          "me"
        ]
      },
      "get": {
        "operationId": "get_me_v1_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Me",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "update_me_v1_me_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Me",
        "tags": [
          "me"
        ]
      }
    },
    "/v1/settings": {
      "get": {
        "operationId": "get_settings_v1_settings_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "update_settings_v1_settings_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/v1/settings/accounts": {
      "get": {
        "operationId": "get_accounts_v1_settings_accounts_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Accounts",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "add_accounts_v1_settings_accounts_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAccountsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Accounts",
        "tags": [
          "settings"
        ]
      }
    },
    "/v1/settings/accounts/import": {
      "post": {
        "operationId": "import_accounts_v1_settings_accounts_import_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportAccountsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportAccountsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Accounts",
        "tags": [
          "settings"
        ]
      }
    },
    "/v1/settings/accounts/verify": {
      "post": {
        "operationId": "verify_accounts_v1_settings_accounts_verify_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyAccountsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Accounts",
        "tags": [
          "settings"
        ]
      }
    },
    "/v1/settings/accounts/{handle}": {
      "delete": {
        "operationId": "delete_account_v1_settings_accounts__handle__delete",
        "parameters": [
          {
            "in": "path",
            "name": "handle",
            "required": true,
            "schema": {
              "title": "Handle",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Account",
        "tags": [
          "settings"
        ]
      }
    }
  }
}